+++ src/BusinessLogic/Tasks/SendDraftTask.php
@@ -8,6 +8,7 @@ use Logeecom\Infrastructure\ServiceRegister;
 use Logeecom\Infrastructure\TaskExecution\Task;
 use Packlink\BusinessLogic\Http\Proxy;
 use Packlink\BusinessLogic\Order\OrderService;
+use Packlink\BusinessLogic\OrderShipmentDetails\OrderShipmentDetailsService;
 
 /**
  * Class UploadDraftTask
@@ -33,6 +34,12 @@ class SendDraftTask extends Task
      * @var Proxy
      */
     private $proxy;
+    /**
+     * OrderShipmentDetailsService instance.
+     *
+     * @var OrderShipmentDetailsService
+     */
+    private $orderShipmentDetailsService;
 
     /**
      * UploadDraftTask constructor.
@@ -98,9 +105,16 @@ class SendDraftTask extends Task
      * @throws \Logeecom\Infrastructure\Http\Exceptions\HttpCommunicationException
      * @throws \Logeecom\Infrastructure\Http\Exceptions\HttpRequestException
      * @throws \Packlink\BusinessLogic\Order\Exceptions\OrderNotFound
+     * @throws \Logeecom\Infrastructure\ORM\Exceptions\QueryFilterInvalidParamException
      */
     public function execute()
     {
+        if ($this->isDraftCreated($this->orderId)) {
+            Logger::logWarning("Draft for order [{$this->orderId}] has been already created. Task is terminating.");
+            $this->reportProgress(100);
+            return;
+        }
+
         $draft = $this->getOrderService()->prepareDraft($this->orderId);
         $this->reportProgress(35);
 
@@ -114,6 +128,28 @@ class SendDraftTask extends Task
 
         $this->getOrderService()->setReference($this->orderId, $reference);
         $this->reportProgress(100);
+    }
+
+    /**
+     * Checks whether draft has already been created for a particular order.
+     *
+     * @param string $orderId Order id in an integrated system.
+     *
+     * @return boolean Returns TRUE if draft has been created; FALSE otherwise.
+     *
+     * @throws \Logeecom\Infrastructure\ORM\Exceptions\QueryFilterInvalidParamException
+     */
+    private function isDraftCreated($orderId)
+    {
+        $draft = $this->getOrderShipmentDetailsService()->getDetailsByOrderId($orderId);
+
+        if ($draft === null) {
+            return false;
+        }
+
+        $reference = $draft->getReference();
+
+        return !empty($reference);
     }
 
     /**
@@ -142,5 +178,19 @@ class SendDraftTask extends Task
         }
 
         return $this->orderService;
+    }
+
+    /**
+     * Retrieves order-shipment details service.
+     *
+     * @return OrderShipmentDetailsService Service instance.
+     */
+    private function getOrderShipmentDetailsService()
+    {
+        if ($this->orderShipmentDetailsService === null) {
+            $this->orderShipmentDetailsService = ServiceRegister::getService(OrderShipmentDetailsService::CLASS_NAME);
+        }
+
+        return $this->orderShipmentDetailsService;
     }
 }